Migration done. Now all the logic is at studyzone map service.
